\name{censor}
\alias{censor}
\title{Censor helper function}
\description{Helper function to censor frequency vectors}
\usage{censor(V, left=0, right=100)}
\arguments{
  \item{V}{Vector}
  \item{left}{left to censor}
  \item{right}{right to censor}
}
\details{Helper function to censor frequency vectors}

\value{censored frequency vector}
\author{Clem Aeppli}
\seealso{\code{\link{truncate}}}
