\name{consensus.variance}
\alias{consensus.variance}
\title{Approximate variance of consensus}
\description{Calculate approximate variance of Blair and Lacy's consensus Cns}
\usage{consensus.variance(V)}
\arguments{
  \item{V}{Frequency vector}
}
\details{Helper function to calculate approximate variance of Blair and Lacy's (2000) consensus Cns.}

\value{Approximate variance of Blair and Lacy's consensus Cns}
\references{Blair, J., and M. Lacy. 2000. Statistics of Ordinal Variation. Sociological Methods & Research 28 (3): 251-280.}
\author{Clem Aeppli}
\seealso{\code{\link{consensus}}}
