\name{D.variance}
\alias{D.variance}
\title{Approximate variance of Leik's D}
\description{Calculate approximate variance of Leik's D}
\usage{D.variance(V)}
\arguments{
  \item{V}{Frequency vector}
}
\details{Helper function to calculate approximate variance of Leik's (1966) D.}

\value{Approximate variance of D}
\references{Leik, R. (1966) A measure of ordinal consensus, Pacific Sociological Review 9(2):85-90.}
\author{Clem Aeppli}
\seealso{\code{\link{Leik}}}
