\name{disper}
\alias{disper}
\title{Measures distance}
\description{This function measures distance between two frequency vectors}
\usage{disper(A, B, metric=2)}
\arguments{
  \item{A}{Vector}
  \item{B}{Vector}
  \item{metric}{city block metric}
}
\details{Calculates the distance between two frequency vectors.}

\value{measure of distance}
\author{Clem Aeppli}
\seealso{\code{\link{concentration}}}
