\name{dispersion}
\alias{dispersion}
\title{Measures dispersion}
\description{This function measures dispersion in a frequency vector}
\usage{dispersion(V, metric=2)}
\arguments{
  \item{V}{Vector}
  \item{metric}{city block metric}
}
\details{dispersion is 1-concentration}

\value{measure of dispersion}
\author{Clem Aeppli}
\seealso{\code{\link{concentration}}}
