\name{dist}
\alias{dist}
\title{Measures distance}
\description{This function measures distance between two frequency vectors}
\usage{dist(A, B, metric=2)}
\arguments{
  \item{A}{Vector}
  \item{B}{Vector}
  \item{metric}{city block metric}
}
\details{distance between vectors}

\value{measure of distance}
\author{Clem Aeppli}
\seealso{\code{\link{concentration}}}
