\name{l.variance}
\alias{l.variance}
\title{Approximate variance of Blair and Lacy's l}
\description{Calculate approximate variance of Blair and Lacy's (2000) l}
\usage{l.variance(V)}
\arguments{
  \item{V}{Frequency vector}
}
\details{Helper function to calculate approximate variance of Blair and Lacy's (2000) l.}

\value{Approximate variance of Blair and Lacy's l}
\references{Blair, J., and M. Lacy. 2000. Statistics of Ordinal Variation. Sociological Methods & Research 28 (3): 251-280.}
\author{Clem Aeppli}
\seealso{\code{\link{BlairLacy}}}
