\name{lsquared.variance}
\alias{lsquared.variance}
\title{Approximate variance of Blair and Lacy's lsquared}
\description{Calculate approximate variance of Blair and Lacy's (2000) lsquared}
\usage{lsquared.variance(V)}
\arguments{
  \item{V}{Frequency vector}
}
\details{Helper function to calculate approximate variance of Blair and Lacy's (2000) lsquared.}

\value{Approximate variance of Blair and Lacy's lsquared}
\references{Blair, J., and M. Lacy. 2000. Statistics of Ordinal Variation. Sociological Methods & Research 28 (3): 251-280.}
\author{Clem Aeppli}
\seealso{\code{\link{BlairLacy}}}
