\name{minnz}
\alias{minnz}
\title{Non-zero minimum}
\description{Helper function to calculate the smallest value of the vector except for 0 (non-zero minimum).}
\usage{minnz(V)}
\arguments{\item{V}{A vector}}
\details{This is a helper function to calculate the non-zero minimum of a vector. The result is the smallest value of the vector, but cannot be zero.}
\value{The function returns the non-zero minimum}
\author{Didier Ruedin}
