\name{patternAgreement}
\alias{patternAgreement}
\title{Calculates patterns agreement}
\description{Helper function to calculate agreement A from a pattern vector.}
\usage{patternAgreement(P, old=FALSE)}
\arguments{
\item{P}{A pattern vector}
\item{old}{Optional argument if the old algorithm for agreement A is to be used. There is normally no reason to set the \code{old} argument.}
}
\details{This is a helper function to calculate agreement A from a pattern vector.}
\value{The function returns the measure of agreement A}
\author{Didier Ruedin}
\seealso{\code{\link{agreement}}}
