\name{Reardon}
\alias{Reardon}
\title{Reardon's entropy}
\description{This function calculates Reardon's (2009) entropy from a frequency vector}
\usage{Reardon(V)}
\arguments{
  \item{V}{Frequency vector}
}
\details{Calculate Reardon's (2009) entropy.
}
\value{measure of entropy}
\references{Reardon, S. 2009. Measures of Ordinal Segregation. Research on Economic Inequality 17:129-55. <DOI: 10.1108/S1049-2585(2009)0000017011>}
\author{Clem Aeppli}
\seealso{\code{\link{disper}}}
