\name{reduceVector}
\alias{reduceVector}
\title{Remove zeros and repeated values}
\description{This is a helper function to remove all zeros and repeated values from a vector.}
\usage{reduceVector(X)}
\arguments{\item{X}{A (frequency) vector}}
\details{This is a helper function to strip all zeros and repeated values from a vector.}
\value{The function returns vector}
\author{Didier Ruedin}
\seealso{\code{\link{agreement}}}
