\name{sd.variance}
\alias{sd.variance}
\title{Approximate variance of the categorical standard deviation}
\description{Calculate approximate variance of the categorical standard deviation}
\usage{sd.variance(V)}
\arguments{
  \item{V}{Frequency vector}
}
\details{Helper function to calculate approximate variance of the categorical standard deviation.}

\value{Approximate variance of the categorical standard deviation}
\author{Clem Aeppli}
\seealso{\code{\link{truncate}}}
