\name{truncatevector}
\alias{truncatevector}
\title{Truncate helper function}
\description{Helper function to truncate frequency vectors}
\usage{truncatevector(V, left=0, right=100)}
\arguments{
  \item{V}{Vector}
  \item{left}{left to truncate}
  \item{right}{right to truncate}
}
\details{Helper function to truncate frequency vectors}

\value{truncated frequency vector}
\author{Clem Aeppli}
\seealso{\code{\link{censor}}}
