\name{var.variance}
\alias{var.variance}
\title{Approximate variance of the consensus (Cns) estimator}
\description{Calculate approximate variance of the consensus (Cns) estimator}
\usage{var.variance(V)}
\arguments{
  \item{V}{Frequency vector}
}
\details{Helper function to calculate approximate variance of the Consensus (Cns) estimator.}

\value{Approximate variance of Blair and Lacy's l}
\references{Blair, J., and M. Lacy. 2000. Statistics of Ordinal Variation. Sociological Methods \& Research 28 (3): 251-280.}
\author{Clem Aeppli}
\seealso{\code{\link{consensus}}}
