\name{agrmt-package}
\alias{agrmt-package}
\alias{agrmt}
\docType{package}
\title{Calculates agreement A}
\description{This package calculates agreement in ordered rating scales. It implements van der Eijk's (2001) measure of agreement A, which can be used to describe agreement, consensus, or polarization among respondents.}
\details{The main functions in this package are \code{\link{agreement}} to calculate agreement "A", and  \code{\link{polarization}} to calculate a polarization score based on agreement A. The package also includes functions to classify distributions according to Galtung's (1969) AJUS-system, and changes over time according to Galtung's (1969) ISD-system. Moreover, the function \code{\link{modes}} can identify the position of multiple modes.}
\author{Didier Ruedin

Maintainer: Didier Ruedin <didier.ruedin@wolfson.oxon.org>
}
\references{van der Eijk, C. (2001) Measuring agreement in ordered rating scales, Quality and Quantity 35(3):325-341.

Galtung, J. (1969) Theory and Methods of Social Research. Oslo: Universitetsforlaget.}
\keyword{ package }
