\name{MRQ}
\alias{MRQ}
\title{Calculates MRQ polarization index}
\description{This function calculates the MRQ polarization index from a population vector.}
\usage{MRQ(Z)}
\arguments{ \item{Z}{Population vector}}
\details{This function implements the polarization index introduced by Montalvo
and Reynal-Querol (2005). Specify a frequency vector.}
\value{Polarization}
\author{Didier Ruedin}
\references{Montalvo, J., and Reynal-Querol, M. 2005. "Ethnic Diversity and Economic Development." Journal of Development Economics 76:293-323.}
